refactor(019/extract-entry): move BinaryScan + entry conversion to entry.rs
Extracts the binary-scan-result-to-PackageDbEntry conversion code
from binary/mod.rs into a new sibling binary/entry.rs:

  pub(crate) struct BinaryScan        (visibility unchanged)
  pub(super) fn version_match_to_entry
  pub(super) fn make_file_level_component
  pub(super) fn note_package_to_entry
  impl PackageDbEntry { ... }         (with_sha256_placeholder helper)
  + 13 inline tests (10 note_package_* + 2 make_file_level_component_* +
    fake_binary_scan helper)

binary/mod.rs:
- declares `mod entry;`
- imports the 4 items via `use entry::{...}`
- drops now-unused imports (ContentHash, Purl, Sha256, Digest,
  rpm_vendor_from_id) — all moved to entry.rs

entry.rs needs `use super::version_strings;` because BinaryScan
references `version_strings::EmbeddedVersionMatch` in
version_match_to_entry's signature.

Net: binary/mod.rs 1510 → 972 LOC (-538). entry.rs is 580 LOC
(290 production + 290 tests).

Verified:
- cargo +stable check --workspace --tests passes clean
- 12 entry tests pass
- All 27 byte-identity goldens unchanged: regen produces zero diff

Per quickstart.md, this is the third of 4 atomic extraction
commits and MUST come before scan.rs (which imports BinaryScan
from entry.rs). Next: scan.rs.
